fix: expose editorId in TuiEditorOptions interface
Fixes #4
