+++ libs/tui/src/index.ts
@@ -2,3 +2,4 @@ export * from './lib/tui-editor.service';
 export * from './lib/tui-editor.component';
 export * from './lib/tui.module';
 export { highlightjsGraphql } from './lib/highlightjs.graphql';
+export { TuiEditorOptions } from './lib/tui-editor.options';
